Fix multiple definition of function pointers in OpenGL ES (Android)
